refactor: TestsBase
